Fix buffer cleanup
It shouldn’t be necessary to jump through so many hoops to gain access to
`org-capture-current-plist' in `org-capture-after-finalize', but
`org-capture-finalize' was not thought with parallelisation in mind.

The main problem is that `org-capture-after-finalize' happens *after* the
org-capture has finished, which means that we cannot access the value of
`org-capture-current-plist' anymore.  `org-capture-plist' is used to store the
variable in a global state to initialise `org-capture', but not to finalise
it.  When you’re only dealing with one buffer, everything is fine, but when
you’ve got two parallel captures, the variable is not updated properly.

All it’d take is to set `org-capture-plist' to `org-capture-current-plist' at
the beginning of `org-capture-finalize'.  I’m considering sending a patch
upstream, but I need to make sure that it’s not causing any problem first.